Refactored feature typedef from console.js
